Refactoring: actions/posts
