+++ src/pages/tag.js
@@ -23,7 +23,7 @@ import { values } from 'lodash';
 
 import ApiClient from '../api/client';
 import { API_HOST } from '../config';
-import { resetCreatePostForm, updateCreatePostForm } from '../actions';
+import { resetCreatePostForm, updateCreatePostForm } from '../actions/posts';
 import { addHashtag, setHashtagPosts } from '../actions/hashtags';
 
 import NotFound from './not-found';
