+++ src/store/comments.js
@@ -36,7 +36,7 @@ const clearComments = (comments) => (
 export default function reducer(state = initialState, action) {
   switch (action.type) {
     case a.river.ADD_POST_TO_RIVER:
-    case a.ADD_POST: {
+    case a.posts.ADD_POST: {
       const comments = action.post.post_comments || [];
 
       state = state.set(action.post.id, List(clearComments(comments)));
@@ -47,11 +47,11 @@ export default function reducer(state = initialState, action) {
     case a.river.SET_POSTS_TO_RIVER:
     case a.river.SET_POSTS_TO_LIKES_RIVER:
     case a.river.SET_POSTS_TO_FAVOURITES_RIVER:
-    case a.SET_USER_POSTS:
+    case a.posts.SET_USER_POSTS:
     case a.hashtags.SET_HASHTAG_POSTS:
     case a.schools.SET_SCHOOL_POSTS:
     case a.geotags.SET_GEOTAG_POSTS:
-    case a.SET_RELATED_POSTS: {
+    case a.posts.SET_RELATED_POSTS: {
       action.posts.forEach(post => {
         const postCopy = cloneDeep(post);
 
@@ -67,7 +67,7 @@ export default function reducer(state = initialState, action) {
       break;
     }
 
-    case a.REMOVE_POST: {
+    case a.posts.REMOVE_POST: {
       state = state.delete(action.id);
 
       break;
