+++ src/store/related_posts.js
@@ -23,7 +23,7 @@ const initialState = i.Map({});
 
 export default function reducer(state = initialState, action) {
   switch (action.type) {
-    case a.SET_RELATED_POSTS: {
+    case a.posts.SET_RELATED_POSTS: {
       const posts = action.posts.map(post => post.id);
 
       state = state.set(action.post_id, i.List(posts));
