+++ src/store/user_posts.js
@@ -17,18 +17,18 @@
 */
 import i from 'immutable';
 
-import * as a from '../actions';
+import { posts } from '../actions';
 
 const initialState = i.Map({});
 
 export default function reducer(state = initialState, action) {
   switch (action.type) {
-    case a.SET_USER_POSTS: {
+    case posts.SET_USER_POSTS: {
       state = state.set(action.user_id, i.List(action.posts.map(post => post.id)));
       break;
     }
 
-    case a.REMOVE_POST: {
+    case posts.REMOVE_POST: {
       for (const user_id of state.keys()) {
         const idx = state.get(user_id).findIndex(user_post_id => (user_post_id === action.id));
